Refactor EditDrawer - remove close button - stylize edit drawer - add icons to edit drawer elements - add default shape for created windows by edit drawer
