+++ src/ui/components/BatteryBar.vue
