+++ src/ui/components/Logo/FIROExpandedLogo.vue
