+++ src/ui/components/Logo/FIROLogo.vue
