+++ src/ui/components/Logo/FIROSmallLogo.vue
